# api-docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "YukiBloom 网站的文章、图片、个人内容和站点页面 API。JSON 响应统一使用 code、message、data；成功 code 为 0，失败 code 与 HTTP 状态码一致。",
        "title": "YukiBloom 后端接口",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/media": {
            "get": {
                "description": "需要先通过管理员登录接口登录。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理图片"
                ],
                "summary": "查询图片库",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量，默认 20，最大 100",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "包含 items、page、limit、total 的分页结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "分页参数无效",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            },
            "post": {
                "description": "需要登录和 CSRF 令牌。表单字段 file 接受 JPEG、PNG 或 WebP，文件最大 10 MiB。",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理图片"
                ],
                "summary": "上传图片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录接口返回的 csrfToken",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "图片文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "图片记录",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/media.Item"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "图片格式或请求无效",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "403": {
                        "description": "CSRF 校验失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "413": {
                        "description": "图片超过大小限制",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "上传失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/media/{id}": {
            "delete": {
                "description": "需要登录和 CSRF 令牌。被文章、页面或站点内容引用的图片不能删除。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理图片"
                ],
                "summary": "删除未引用图片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "图片 UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录接口返回的 csrfToken",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功，data 为 null",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "400": {
                        "description": "图片 ID 格式错误",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "403": {
                        "description": "CSRF 校验失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "404": {
                        "description": "图片不存在",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "409": {
                        "description": "图片仍被内容引用",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
...
        "/uploads/{key}": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "图片"
                ],
                "summary": "获取公开图片文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "图片存储文件名",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "图片文件",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "图片不存在",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "读取图片失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apiresponse.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "media.Item": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "mimeType": {
                    "type": "string"
                },
                "sizeBytes": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
...
    }
}